Added ability to load default, global and local settings_dialog
I'm not going to lie to you. I'm pretty dam drunk right now. There's no-one checking my code (there never is in this project), and I don't want to be held responsable for this commit. Whatever... It's probably not any worse than any of other commits that I've made sober.