Added the ability to set transparancy
This solves issue #24.

As for now, it's only possible to set transparancy when PyMuPDF is used. If PyPDF is used, this setting does nothing. I'm not sure if it's possible to implement this with PyPDF as well, I haven't looked into it and I won't do that anytime soon.

To add transparancy I've created an entierly new routine for coloring. Instead of "rg" and "RG" in the byte stream and replace the content with a new color as we usually do, I'm reading all drawings from the pdf and adding them to a new pdf with changed color and opacity.