Removed kicad engine option from CLI.
It's not possible to reload the color templates in KiCad when not using the KiCad GUI.
I've only removed the visible option from the CLI. One can still specify 'kicad' if you want to test it.

Also changed so that kicad engine isn't the default option when using KiCad 7.0, since the color of the drawing sheet (frame) is always the same.